pyproject.toml
[project]
name = "what-where"
version = "0.1.0"
description = ""
authors = [{name = "Eivinas Butkus"}]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel (>=6.29.5,<7.0.0)",
    "matplotlib (>=3.10.0,<4.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "plotly (>=6.0.0,<7.0.0)",
    "pykrige (>=1.7.2,<2.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "statsmodels (>=0.14.4,<0.15.0)",
    "torch (>=2.6.0,<3.0.0)",
    "torchvision (>=0.21.0,<0.22.0)",
    "torchaudio (>=2.6.0,<3.0.0)",
    "hydra-core (>=1.3.2,<2.0.0)",
    "h5py (>=3.13.0,<4.0.0)",
    "tensorboard (>=2.19.0,<3.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "nvtx (>=0.2.11,<0.3.0)",
    "pyarrow (>=20.0.0,<21.0.0)",
    "nbconvert (>=7.16.6,<8.0.0)",
    "kornia (>=0.8.2,<0.9.0)",
    "opencv-python (>=4.11.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
